requirements.txt
python-dateutil
requests
django
djangorestframework
django-registration
django-countries
django-modeltranslation
# mysqlclient
# psycopg2